updated build to include most recent versions of components, updated dependencies and webpack config
