+++ client/App.js
@@ -1,21 +1,35 @@
-/* eslint-disable no-unused-vars */
-import React from "react";
+import React from 'react';
+import './App.css';
+import { useState } from 'react';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
-import Navbar from "./components/Navbar";
-import Routes from "./Routes";
-import Home from "./components/HomePage";
-import SinglePage from "./components/SinglePage";
-import AllBalls from "./components/AllBalls";
+// Views
+import HomeScreen from './views/HomeScreen';
+import ProductScreen from './views/ProductScreen';
+import CartScreen from './views/CartScreen';
 
-const App = () => {
+// Components
+import Navbar from './components/Navbar.js';
+import Backdrop from './components/Backdrop.js';
+import SideDrawer from './components/SideDrawer.js';
+
+function App() {
+  const [sideToggle, setSideToggle] = useState(false);
+
   return (
-    <div>
-      <Navbar />
-
-      <Home />
-      {/* <Routes /> */}
-    </div>
+    <Router>
+      <Navbar click={() => setSideToggle(true)} />
+      <SideDrawer show={sideToggle} click={() => setSideToggle(false)} />
+      <Backdrop show={sideToggle} click={() => setSideToggle(false)} />
+      <main>
+        <Switch>
+          <Route exact path='/' component={HomeScreen} />
+          <Route exact path='/product/:id' component={ProductScreen} />
+          <Route exact path='/cart' component={CartScreen} />
+        </Switch>
+      </main>
+    </Router>
   );
-};
+}
 
 export default App;
