+++ client/components/Backdrop.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Backdrop.css';
+import './componentStyles/Backdrop.css';
 
 const Backdrop = ({ show, click }) => {
   return show && <div className='backdrop' onClick={click}></div>;
