+++ client/testComponents/Routes.js
@@ -1,3 +1,5 @@
+// NOTE: file likely not needed
+
 // import React, { Component, Fragment } from "react";
 // import { connect } from "react-redux";
 // import { withRouter, Route, Switch, Redirect } from "react-router-dom";
