+++ client/testComponents/SinglePage.js
