+++ webpack.config.js
@@ -1,7 +1,5 @@
 module.exports = {
-  entry: [
-    './client/index.js'
-  ],
+  entry: ['./client/index.js'],
   output: {
     path: __dirname,
     filename: './public/bundle.js'
@@ -14,11 +12,13 @@ module.exports = {
         exclude: /node_modules/,
         loader: 'babel-loader',
         options: {
-          presets: [
-            '@babel/preset-react'
-          ]
+          presets: ['@babel/preset-react']
         }
+      },
+      {
+        test: /\.css$/i,
+        use: ['style-loader', 'css-loader']
       }
     ]
   }
-}
+};
